Simplify nested client namespaces
