+++ lib/cisco_node_utils/node.rb
@@ -200,7 +200,7 @@ module Cisco
 
     # "hidden" API - used for UT but shouldn't be used elsewhere
     def connect(*args)
-      @client = Cisco::Client::Client.create(*args)
+      @client = Cisco::Client.create(*args)
       @cmd_ref = CommandReference.new(product:  product_id,
                                       platform: @client.platform,
                                       cli:      @client.supports?(:cli))
